Edited the css, index, modal, script and readme files/added jquery code and bootstrap
Added jQuery and bootStrap code. Changed the classes to bootstrap style. Also added a nav bar to the top of the home page and linked my portfolio page to the home page. Added a link to the pokemon details page from the home page in case more details are needed about pokemons. Added the changes details to the ReadMe page